requirements.txt
yaml~=0.2.5
pyyaml~=6.0
tqdm~=4.64.0
opencv-python~=4.6.0.66
matplotlib~=3.4.3
numpy~=1.23.0
pandas~=1.2.4
seaborn~=0.11.2
pillow~=9.0.1
torchvision~=0.12.0
requests~=2.27.1
flask~=2.0.3
albumentations~=1.0.3